Move to jacobians.py and fix test
