+++ external/fv3fit/fv3fit/train_microphysics.py
@@ -23,7 +23,8 @@ from fv3fit.emulation import models, train, ModelCheckpointCallback
 from fv3fit.emulation.data import TransformConfig, nc_dir_to_tf_dataset
 from fv3fit.emulation.data.config import SliceConfig
 from fv3fit.emulation.layers import ArchitectureConfig
-from fv3fit.emulation.keras import save_model, get_jacobians, standardize_jacobians
+from fv3fit.emulation.jacobian import get_jacobians, standardize_jacobians
+from fv3fit.emulation.keras import save_model
 from fv3fit.emulation.losses import CustomLoss
 from fv3fit.wandb import (
     WandBConfig,
